Refactor report into code into its own package
