+++ src/main/java/uk/gov/hmcts/ethos/replacement/docmosis/reports/casescompleted/CasesCompletedReport.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.ethos.replacement.docmosis.reports;
+package uk.gov.hmcts.ethos.replacement.docmosis.reports.casescompleted;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
