+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/service/ListingServiceTest.java
@@ -14,7 +14,7 @@ import uk.gov.hmcts.ecm.common.model.ccd.types.*;
 import uk.gov.hmcts.ecm.common.model.listing.ListingData;
 import uk.gov.hmcts.ecm.common.model.listing.ListingDetails;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.BFHelperTest;
-import uk.gov.hmcts.ethos.replacement.docmosis.reports.CasesCompletedReport;
+import uk.gov.hmcts.ethos.replacement.docmosis.reports.casescompleted.CasesCompletedReport;
 import uk.gov.hmcts.ethos.replacement.docmosis.utils.InternalException;
 
 import java.io.IOException;
